Feature: new method `addMatrixConsIndicator` (#1024)
* ENH: new method `addMatrixConsIndicator`

similar to `addConsIndicator` but for matrix

* Update CHANGELOG.md

* Update test_matrix_variable.py

* use pyscipopt code and doc style

Condensed the addMatrixConsIndicator method signature to a single line and updated the docstring for improved clarity and consistency. Parameter descriptions were streamlined and default values were explicitly stated.

* use pyscipopt code and doc style

* Add return type to add_matrix_constraint method

Specified the return type of the Model.add_matrix_constraint method as MatrixConstraint to improve type clarity and support better static analysis.

* Apply suggestions from code review

Co-authored-by: João Dionísio <57299939+Joao-Dionisio@users.noreply.github.com>

---------

Co-authored-by: João Dionísio <57299939+Joao-Dionisio@users.noreply.github.com>